refactor: Rename `scoring.ts`
