+++ src/app/tournament/[uuid]/sessions/_components/ViewGraphModal.tsx
@@ -1,7 +1,7 @@
 import { useMemo } from "react";
 import { BarChart } from "@/elements/BarChart";
 import { Modal } from "@/elements/Modal";
-import { rankPlayers } from "@/lib/scoring";
+import { rankPlayers } from "@/lib/scores";
 import type { Player } from "@/types/app.types";
 
 type ViewGraphModalProps = {
