+++ src/lib/scores.ts
